langchain-langgraph/pyproject.toml
[project]
name = "langchain-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain[openai]>=0.3.25",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
